creating json-viewer.tsx and adjustments
